@@ -12,8 +12,8 @@ import ti.modules.titanium.geolocation.android.LocationProviderProxy.LocationPro
 import android.annotation.SuppressLint;
 import android.content.Context;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.GoogleApiAvailability;
@@ -76,6 +76,7 @@ public class FusedLocationProvider
 		}
 		try {
 			Class.forName("com.google.android.gms.common.GoogleApiAvailability");
+			Class.forName("com.google.android.gms.location.FusedLocationProviderClient");
 			return PlayServices.validVersion() && PlayServices.available(context);
 		} catch (Exception e) {
 			useFusedLocation = false;
