@@ -211,7 +211,13 @@ AndroidModuleBuilder.prototype.validate = function validate(logger, config, cli)
 				process.exit(1);
 			}
 
-			const targetSDKMap = {};
+			const targetSDKMap = {
+
+				// placeholder for gradle to use
+				[this.compileSdkVersion]: {
+					sdk: this.compileSdkVersion
+				}
+			};
 			Object.keys(this.androidInfo.targets).forEach(function (id) {
 				var t = this.androidInfo.targets[id];
 				if (t.type === 'platform') {
@@ -221,69 +227,17 @@ AndroidModuleBuilder.prototype.validate = function validate(logger, config, cli)
 
 			// check the Android SDK we require to build exists
 			this.androidCompileSDK = targetSDKMap[this.compileSdkVersion];
-			if (!this.androidCompileSDK) {
-				logger.error(__('Unable to find Android SDK API %s', this.compileSdkVersion));
-				logger.error(__('Android SDK API %s is required to build Android modules', this.compileSdkVersion) + '\n');
-				process.exit(1);
-			}
 
 			// if no target sdk, then default to most recent supported/installed
 			if (!this.targetSDK) {
-				const levels = Object.keys(targetSDKMap).sort();
-
-				for (let i = levels.length - 1; i >= 0; i--) {
-					if (levels[i] >= this.minSupportedApiLevel && levels[i] <= this.maxSupportedApiLevel) {
-						this.targetSDK = levels[i];
-						break;
-					}
-				}
-
-				if (!this.targetSDK) {
-					logger.error(__('Unable to find a suitable installed Android SDK that is >=%s and <=%s', this.minSupportedApiLevel, this.maxSupportedApiLevel) + '\n');
-					process.exit(1);
-				}
+				this.targetSDK = this.maxSupportedApiLevel;
 			}
-
-			// check that we have this target sdk installed
 			this.androidTargetSDK = targetSDKMap[this.targetSDK];
 
 			if (!this.androidTargetSDK) {
-				logger.error(__('Target Android SDK %s is not installed', this.targetSDK) + '\n');
-
-				const sdks = Object.keys(targetSDKMap).filter(function (ver) {
-					return ver > this.minSupportedApiLevel;
-				}.bind(this)).sort().filter(function (s) { return s >= this.minSDK; }, this);
-
-				if (sdks.length) {
-					logger.log(__('To target Android SDK %s, you first must install it using the Android SDK manager.', String(this.targetSDK).cyan) + '\n');
-					logger.log(
-						appc.string.wrap(
-							__('Alternatively, you can set the %s in the %s section of the tiapp.xml to one of the following installed Android target SDKs: %s', '<uses-sdk>'.cyan, '<android> <manifest>'.cyan, sdks.join(', ').cyan),
-							config.get('cli.width', 100)
-						)
-					);
-					logger.log();
-					logger.log('<ti:app xmlns:ti="http://ti.appcelerator.org">'.grey);
-					logger.log('    <android>'.grey);
-					logger.log('        <manifest>'.grey);
-					logger.log(('            <uses-sdk '
-						+ (this.minSDK ? 'android:minSdkVersion="' + this.minSDK + '" ' : '')
-						+ 'android:targetSdkVersion="' + sdks[0] + '" '
-						+ (this.maxSDK ? 'android:maxSdkVersion="' + this.maxSDK + '" ' : '')
-						+ '/>').magenta);
-					logger.log('        </manifest>'.grey);
-					logger.log('    </android>'.grey);
-					logger.log('</ti:app>'.grey);
-					logger.log();
-				} else {
-					logger.log(__('To target Android SDK %s, you first must install it using the Android SDK manager', String(this.targetSDK).cyan) + '\n');
-				}
-				process.exit(1);
-			}
-
-			if (!this.androidTargetSDK.androidJar) {
-				logger.error(__('Target Android SDK %s is missing "android.jar"', this.targetSDK) + '\n');
-				process.exit(1);
+				this.androidTargetSDK = {
+					sdk: this.targetSDK
+				};
 			}
 
 			if (this.targetSDK < this.minSDK) {
@@ -300,6 +254,9 @@ AndroidModuleBuilder.prototype.validate = function validate(logger, config, cli)
 				// print warning that version this.targetSDK is not tested
 				logger.warn(__('Building with Android SDK %s which hasn\'t been tested against Titanium SDK %s', ('' + this.targetSDK).cyan, this.titaniumSdkVersion));
 			}
+
+			// get javac params
+			this.javacMaxMemory = cli.timodule.properties['android.javac.maxmemory'] && cli.timodule.properties['android.javac.maxmemory'].value || config.get('android.javac.maxMemory', '3072M');
 
 			// detect java development kit
 			appc.jdk.detect(config, null, function (jdkInfo) {
@@ -534,10 +491,17 @@ AndroidModuleBuilder.prototype.generateRootProjectFiles = async function generat
 	this.logger.info(__('Generating root project files'));
 
 	// Copy our SDK's gradle files to the build directory. (Includes "gradlew" scripts and "gradle" directory tree.)
-	// The below install method will also generate a "gradle.properties" file.
 	const gradlew = new GradleWrapper(this.buildDir);
 	gradlew.logger = this.logger;
 	await gradlew.installTemplate(path.join(this.platformPath, 'templates', 'gradle'));
+
+	// Create a "gradle.properties" file. Will add network proxy settings if needed.
+	const gradleProperties = await gradlew.fetchDefaultGradleProperties();
+	gradleProperties.push({
+		key: 'org.gradle.jvmargs',
+		value: `-Xmx${this.javacMaxMemory} -Dkotlin.daemon.jvm.options="-Xmx${this.javacMaxMemory}"`
+	});
+	await gradlew.writeGradlePropertiesFile(gradleProperties);
 
 	// Create a "local.properties" file providing a path to the Android SDK/NDK directories.
 	await gradlew.writeLocalPropertiesFile(this.androidInfo.sdk.path, this.androidInfo.ndk.path);
@@ -603,6 +567,7 @@ AndroidModuleBuilder.prototype.generateModuleProject = async function generateMo
 		moduleMavenArtifactId: mavenArtifactId,
 		moduleName: this.manifest.name,
 		moduleVersion: this.manifest.version,
+		moduleArchitectures: this.manifest.architectures.split(' '),
 		tiBindingsJsonPath: path.join(this.platformPath, 'titanium.bindings.json'),
 		tiMavenUrl: encodeURI('file://' + path.join(this.platformPath, 'm2repository').replace(/\\/g, '/')),
 		tiSdkModuleTemplateDir: this.moduleTemplateDir,
